+++ part_1/askisi_4.cpp
@@ -13,8 +13,8 @@ typedef struct record
     string commodity;
     string transport_mode;
     string measure;
-    long int value;
-    long int cumulative;
+    long long value;
+    long long cumulative;
 } Record;
 
 int binaryInterpolationSearchDates(vector<Record> &data, int key, int left, int right);
