+++ part_2/askisi_2.cpp
@@ -12,8 +12,8 @@ typedef struct record
     string commodity;
     string transport_mode;
     string measure;
-    long int value;
-    long int cumulative;
+    long long value;
+    long long cumulative;
 } Record;
 
 typedef struct BstNode
@@ -25,7 +25,7 @@ typedef struct BstNode
 
 BstNode *getNewNode(Record element);
 BstNode *insertValue(BstNode *root, Record element);
-string searchValue(BstNode *root, long int value);
+string searchValue(BstNode *root, long long value);
 void inOrderRepr(BstNode *node);
 void saveNodesInOrder(BstNode *node, vector<BstNode *> &nodes);
 BstNode *sortedVectorToBST(vector<BstNode *> &nodes, int first, int last);
@@ -33,8 +33,8 @@ BstNode *makeBalancedTree(BstNode *root);
 
 void menuBSTValue(BstNode *root);
 BstNode *printInOrder(BstNode *root);
-long int findMin(BstNode *root);
-long int findMax(BstNode *root);
+long long findMin(BstNode *root);
+long long findMax(BstNode *root);
 BstNode *printSearchMin(BstNode *root);
 BstNode *printSearchMax(BstNode *root);
 
@@ -108,7 +108,7 @@ BstNode *insertValue(BstNode *root, Record element)
     return root;
 }
 
-string searchValue(BstNode *root, long int value)
+string searchValue(BstNode *root, long long value)
 {
     if (root == NULL)
     {
@@ -239,7 +239,7 @@ BstNode *printSearchMin(BstNode *root)
 {
     // Code For Searching Date by Minimum Value
 
-    long int min = findMin(root);
+    long long min = findMin(root);
     cout << "Minimum value is " << min << endl;
     cout << "Corresponding Date is: " << searchValue(root, min) << endl;
     return root;
@@ -249,7 +249,7 @@ BstNode *printSearchMax(BstNode *root)
 {
     // Code For Searching Date by Maximum Value
 
-    long int max = findMax(root);
+    long long max = findMax(root);
     cout << "Maximum value is " << max << endl;
     cout << "Corresponding Date is: " << searchValue(root, max) << endl;
     return root;
@@ -268,7 +268,7 @@ BstNode *printInOrder(BstNode *root)
 }
 
 // Function that returns the minimum value of the BST tree
-long int findMin(BstNode *root)
+long long findMin(BstNode *root)
 {
 
     vector<BstNode *> sortednodes; // vector to store sorted values after inorder traversal
@@ -277,7 +277,7 @@ long int findMin(BstNode *root)
 }
 
 // Function that returns the maximum value of the BST tree
-long int findMax(BstNode *root)
+long long findMax(BstNode *root)
 {
 
     vector<BstNode *> sortednodes; // vector to store sorted values after inorder traversal
